# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraBook API",
        "description": "LaraBook API",
        "contact": {
            "email": "hubertlipinskipl@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Localhost server"
        }
    ],
    "paths": {
        "/auth/logout": {
            "put": {
                "tags": [
                    "auth"
                ],
                "summary": "Log out",
                "description": "Log out",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "example": "Successfully logged out"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Authenticate to API",
                "in": "header",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                },
                "scheme": "token"
            }
        }
    },
    "security": [
        []
    ]
}